pyproject.toml
[tool.poetry]
package-mode = false

[tool.poetry.dependencies]
python = "3.13.11"
fastapi = "^0.128.0"
uvicorn = { version = "^0.40.0", extras = ["standard"] }
pydantic = { extras = ["email"], version = "^2.12.5" }
pydantic-settings = "^2.12.0"
sqlalchemy = { extras = ["asyncio"], version = "^2.0.45" }
asyncpg = "^0.31.0"
alembic = "^1.18.1"
passlib = "1.7.4"
bcrypt = "4.0.1"
pyjwt = {extras = ["crypto"], version = "^2.12.1"}
password-validator = "^1.0"

[tool.poetry.group.dev.dependencies]
pytest = "^9.0.2"
pytest-asyncio = "^1.3.0"
httpx = "^0.28.1"
